Display: Move core counters/stats to HW file.
This separates things better, so not everything is pulling in HLE.
