+++ Core/Debugger/WebSocket/GPUStatsSubscriber.cpp
@@ -18,7 +18,6 @@
 #include <mutex>
 #include <vector>
 #include "Core/Debugger/WebSocket/GPUStatsSubscriber.h"
-#include "Core/HLE/sceDisplay.h"
 #include "Core/HW/Display.h"
 #include "Core/System.h"
