+++ Core/Reporting.cpp
@@ -40,8 +40,8 @@
 #include "Core/FileSystems/BlockDevices.h"
 #include "Core/FileSystems/MetaFileSystem.h"
 #include "Core/HLE/Plugins.h"
-#include "Core/HLE/sceDisplay.h"
 #include "Core/HLE/sceKernelMemory.h"
+#include "Core/HW/Display.h"
 #include "Core/ELF/ParamSFO.h"
 #include "GPU/GPUInterface.h"
 #include "GPU/GPUState.h"
