+++ UI/EmuScreen.cpp
@@ -63,7 +63,6 @@ using namespace std::placeholders;
 #include "GPU/Vulkan/DebugVisVulkan.h"
 #endif
 #include "Core/HLE/sceCtrl.h"
-#include "Core/HLE/sceDisplay.h"
 #include "Core/HLE/sceSas.h"
 #include "Core/Debugger/SymbolMap.h"
 #include "Core/SaveState.h"
